requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-camel-case==1.2.0
feedparser==6.0.1
flake8==3.8.4
idna==2.10
importlib-metadata==2.0.0
isort==5.5.4
kombu==5.0.2
mccabe==0.6.1
multidict==4.7.6
prompt-toolkit==3.0.7
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
sgmllib3k==1.0.0
six==1.15.0
sqlparse==0.4.0
typing-extensions==3.7.4.3
vcrpy==4.1.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.0
zipp==3.3.0
